requirements.txt
langchain==0.0.138
python-dotenv==1.0.0
openai==0.27.4
pymupdf==1.21.1
werkzeug==2.2.2
tiktoken==0.3.3
jsonify==0.5
chromadb==0.3.21
prompt_toolkit
promptlayer
flask[async]